Devtooling 194 (#768)
* changes made

* Refactored line base settings

* added comment

* used setnillable value
